refactor: move output package to cli/output
Move internal/output to internal/cli/output to better organize
CLI-specific packages. This is part of the clean architecture
refactoring to separate CLI concerns from core business logic.
