+++ internal/cli/commands/param/list/list.go
@@ -12,8 +12,8 @@ import (
 	"github.com/urfave/cli/v3"
 
 	"github.com/mpyw/suve/internal/api/paramapi"
+	"github.com/mpyw/suve/internal/cli/output"
 	"github.com/mpyw/suve/internal/infra"
-	"github.com/mpyw/suve/internal/output"
 	"github.com/mpyw/suve/internal/parallel"
 )
