+++ internal/cli/commands/param/show/show.go
@@ -11,10 +11,10 @@ import (
 	"github.com/urfave/cli/v3"
 
 	"github.com/mpyw/suve/internal/api/paramapi"
+	"github.com/mpyw/suve/internal/cli/output"
 	"github.com/mpyw/suve/internal/cli/pager"
 	"github.com/mpyw/suve/internal/infra"
 	"github.com/mpyw/suve/internal/jsonutil"
-	"github.com/mpyw/suve/internal/output"
 	"github.com/mpyw/suve/internal/timeutil"
 	"github.com/mpyw/suve/internal/version/paramversion"
 )
