+++ internal/staging/runner/apply.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/mpyw/suve/internal/cli/output"
 	"github.com/mpyw/suve/internal/maputil"
-	"github.com/mpyw/suve/internal/output"
 	"github.com/mpyw/suve/internal/parallel"
 	"github.com/mpyw/suve/internal/staging"
 )
